# openapi-schema.yml
openapi: 3.0.2
info:
  title: ''
  version: ''
paths:
  /laticreteapp/productLine/:
    get:
      operationId: listproductLines
      description: ''
      parameters:
      - name: search
        required: false
        in: query
        description: A search term.
        schema:
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items:
                  properties:
                    lineName:
                      type: string
                    image:
                      type: string
                      format: binary
                    details:
                      type: string
                      nullable: true
                      maxLength: 1000
                  required:
                  - lineName
                  - image
          description: ''
    post:
      operationId: CreateproductLine
      description: ''
      parameters: []
      requestBody:
        content:
          application/json:
            schema: &id001
              properties:
                lineName:
                  type: string
                image:
                  type: string
                  format: binary
                details:
                  type: string
                  nullable: true
                  maxLength: 1000
              required:
              - lineName
              - image
          application/x-www-form-urlencoded:
            schema: *id001
          multipart/form-data:
            schema: *id001
      responses:
        '200':
          content:
            application/json:
              schema:
                properties:
                  lineName:
                    type: string
                  image:
                    type: string
                    format: binary
                  details:
                    type: string
                    nullable: true
                    maxLength: 1000
                required:
                - lineName
                - image
          description: ''
...